rename instr_handler to pip_handler and add instruments folder
