+++ api/src/opentrons/hardware_control/__init__.py
@@ -14,8 +14,6 @@ from .adapters import SynchronousAdapter
 from .api import API
 from .pause_manager import PauseManager
 from .backends import Controller, Simulator
-from .pipette import Pipette
-from .gripper import Gripper
 from .types import (
     CriticalPoint,
     NoTipAttachedError,
@@ -28,7 +26,7 @@ from .thread_manager import ThreadManager
 from .execution_manager import ExecutionManager
 from .threaded_async_lock import ThreadedAsyncLock, ThreadedAsyncForbidden
 from .protocols import HardwareControlAPI
-from .instrument_abc import AbstractInstrument
+from .instruments import AbstractInstrument, Pipette, Gripper
 
 ThreadManagedHardware = ThreadManager[HardwareControlAPI]
 SyncHardwareAPI = SynchronousAdapter[HardwareControlAPI]
