+++ api/src/opentrons/hardware_control/instruments/__init__.py
@@ -1,0 +1,6 @@
+from .instrument_abc import AbstractInstrument
+from .pipette import Pipette
+from .gripper import Gripper
+
+
+__all__ = ["AbstractInstrument", "Pipette", "Gripper"]
