+++ api/src/opentrons/hardware_control/protocols/instrument_configurer.py
@@ -4,7 +4,7 @@ from typing_extensions import Protocol
 from opentrons_shared_data.pipette.dev_types import PipetteName
 from opentrons.types import Mount
 
-from ..pipette import Pipette
+from opentrons.hardware_control.instruments.pipette import Pipette
 from ..dev_types import PipetteDict
 from ..types import CriticalPoint
